\name{unsystation-package}
\alias{unsystation-package}
\alias{unsystation}
\docType{package}

\title{
A second-order stationarity of time series based on unsystematic sub-samples
}

\description{
The package implements a new method for testing the stationarity of time series, where the test statistic is obtained from measuring and maximising the difference in the second-order structure over pairs of randomly drawn intervals.
}

\details{
\tabular{ll}{
Package: \tab unsystation\cr
Type: \tab Package\cr
Version: \tab 0.2.0\cr
Date: \tab 2018-05-23\cr
License: \tab GPL (>= 2)\cr
}

The main routine of the package is \code{\link{unsys.station.test}}.

}
\author{
Haeran Cho

Maintainer: Haeran Cho <haeran.cho@bristol.ac.uk>
}

\references{
H. Cho (2016) A second-order stationarity of time series based on unsystematic sub-samples. Stat, vol. 5, 262-277.
}

\keyword{stationarity test, unsystematic sampling, local stationarity}

